backend refactor to promote testability
A lot of these changes were made to help map the Perl package model into the
Java package model.

In Perl, a "package" is named, for example, Devel::Cover.  In Java, the
package would be "Devel" and the class would be "Cover".  Many tools and file
formats based on Java tech presume this sort of hierarchy.  Thus, the backend
was rewritten to take this into account for the Clover xml file format.